Refactor SpringBootJNDIDataSourceTest (#15250)
* Refactor SpringBootJNDIDataSourceTest

* Refactor SpringBootJNDIDataSourceTest

* Refactor SpringBootJNDIDataSourceTest

* Refactor SpringBootJNDIDataSourceTest
