+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-boot-starter/src/test/java/org/apache/shardingsphere/spring/boot/jndi/SpringBootJNDIDataSourceTest.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.spring.boot;
+package org.apache.shardingsphere.spring.boot.jndi;
 
 import org.apache.shardingsphere.driver.jdbc.core.datasource.ShardingSphereDataSource;
 import org.apache.shardingsphere.infra.database.DefaultSchema;
-import org.apache.shardingsphere.spring.boot.fixture.TestJndiInitialContextFactory;
+import org.apache.shardingsphere.spring.boot.jndi.fixture.InitialDataSourceInitialContextFactory;
 import org.apache.shardingsphere.test.mock.MockedDataSource;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -49,16 +49,16 @@ public class SpringBootJNDIDataSourceTest {
     
     @BeforeClass
     public static void setUp() {
-        System.setProperty(Context.INITIAL_CONTEXT_FACTORY, TestJndiInitialContextFactory.class.getName());
-        TestJndiInitialContextFactory.bind("java:comp/env/jdbc/jndi0", new MockedDataSource());
-        TestJndiInitialContextFactory.bind("java:comp/env/jdbc/jndi1", new MockedDataSource());
+        System.setProperty(Context.INITIAL_CONTEXT_FACTORY, InitialDataSourceInitialContextFactory.class.getName());
+        InitialDataSourceInitialContextFactory.bind("java:comp/env/jdbc/ds0", new MockedDataSource());
+        InitialDataSourceInitialContextFactory.bind("java:comp/env/jdbc/ds1", new MockedDataSource());
     }
     
     @Test
     public void assertDataSources() {
         Map<String, DataSource> dataSources = dataSource.getContextManager().getMetaDataContexts().getMetaData(DefaultSchema.LOGIC_NAME).getResource().getDataSources();
         assertThat(dataSources.size(), is(2));
-        assertTrue(dataSources.containsKey("jndi0"));
-        assertTrue(dataSources.containsKey("jndi1"));
+        assertTrue(dataSources.containsKey("ds0"));
+        assertTrue(dataSources.containsKey("ds1"));
     }
 }
