+++ shardingsphere-jdbc/shardingsphere-jdbc-spring/shardingsphere-jdbc-core-spring/shardingsphere-jdbc-core-spring-boot-starter/src/test/resources/application-jndi.properties
@@ -15,10 +15,6 @@
 # limitations under the License.
 #
 
-spring.shardingsphere.datasource.names=jndi0,jndi1
-spring.shardingsphere.datasource.jndi0.jndi-name=java:comp/env/jdbc/jndi0
-spring.shardingsphere.datasource.jndi1.jndi-name=jdbc/jndi1
-
-spring.shardingsphere.rules.readwrite-splitting.data-sources.readwrite_ds.type=Static
-spring.shardingsphere.rules.readwrite-splitting.data-sources.readwrite_ds.props.write-data-source-name=jndi0
-spring.shardingsphere.rules.readwrite-splitting.data-sources.readwrite_ds.props.read-data-source-names=jndi1
+spring.shardingsphere.datasource.names=ds0,ds1
+spring.shardingsphere.datasource.ds0.jndi-name=java:comp/env/jdbc/ds0
+spring.shardingsphere.datasource.ds1.jndi-name=jdbc/ds1
